move sign out button to main nav
