+++ src/routes/account/+page.svelte
@@ -178,11 +178,6 @@
 			loading = false;
 		}
 	};
-
-	const signOut = async () => {
-		await supabase.auth.signOut();
-		goto('/');
-	};
 </script>
 
 <div class="profile">
@@ -240,9 +235,6 @@
 			/>
 			<sl-button type="submit" class="update" aria-live="polite" {loading}>
 				<span>Update Profile</span>
-			</sl-button>
-			<sl-button type="button" variant="warning" aria-live="polite" on:click={signOut}>
-				<span>Sign Out</span>
 			</sl-button>
 		</form>
 	</section>
